Fix ModuleNotFoundError due to missing pkg_resources
The `ModuleNotFoundError: pkg_resources` occurs when the underlying
virtualenv does not have `setuptools` installed, e.g. when installed
with `uv tool install bibdesk2zotero`. This is because `pybtex<=0.24`
import `pkg_resources` module, which is contained in the `setuptools`
package. Note that for `pybtex>0.24` (not released yet), they drop
dependency on it and use instead `importlib`, part of the python
stdlib. This means `setuptools` will not be needed for `pybtex>0.24`.

I did a bisect search and found that setuptools>=40.4.3 is required for
the bibdesk2zotero script to work.